requirement.txt
pyqtgraph == 0.11.0
pyserial
PyOpenGL
joblib